+++ src/simple_palette_utils/__init__.py
@@ -2,5 +2,4 @@ from ._palettes import X11_COLORS, DARK_COLORS, LIGHT_COLORS, COLORBLIND8_COLORS
 from ._palettes import COLOR_LISTS, COLOR_LIST_X11, COLOR_LIST_LIGHT, COLOR_LIST_DARK, COLOR_LIST_COLORBLIND8, COLOR_LIST_COLORBLIND12, COLOR_LIST_COLORBLIND15, COLOR_LIST_COLORBLIND24
 from ._palettes import colors, x11_colors, parse_rgb, parse_rgba, ColorProvider, color_lists, add_color_list
 from ._palettes import PALETTES, PALETTE_AUTO, PALETTE_GRAYSCALE, PALETTE_X11, PALETTE_LIGHT, PALETTE_DARK, PALETTE_COLORBLIND8, PALETTE_COLORBLIND12, PALETTE_COLORBLIND15, PALETTE_COLORBLIND24
-from ._palettes import default_palette, fill_palette, grayscale_palette, generate_palette_list, add_palette
-from ._palettes import palette_list_to_triplets
+from ._palettes import palettes, default_palette, fill_palette, grayscale_palette, generate_palette_list, add_palette, palette_list_to_triplets